% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_indicator.R
\name{analyse_indicator}
\alias{analyse_indicator}
\title{Complete  analysis for one hypothesis}
\usage{
analyse_indicator(data, dependent.var, independent.var = NULL,
  hypothesis.type, sampling.strategy.cluster = FALSE,
  sampling.strategy.stratified = FALSE,
  do.for.each.unique.value.in.var = NULL)
}
\arguments{
\item{data}{}

\item{dependent.var}{string with the column name in `data` of the dependent variable}

\item{hypothesis.type}{the type of hypothesis as a string. Allowed values are "direct_reporting", "group_difference", "limit", "correlation" or "change"}

\item{sampling.strategy.cluster}{set to TRUE if you used cluster sampling}

\item{sampling.strategy.stratified}{set to TRUE if you used stratified sampling}

\item{do.for.each.unique.value.in.var}{if you want to repeat the analysis for multiple subsets of the data, specify the column name in `data` by which to split the dataset}

\item{independen.var}{string with the column name in `data` of the independent variable}
}
\value{
A list with 1. the summary.statistic, 2. the hypothesis.test.result, and 3. the visualisation as a ggplot object
}
\description{
Produce summary statistics, hypothesis tests and plot objects for a hypothesis
}
\details{
this function takes the data, information about your variables of interest, hypothesis type and sampling strategy. It selects the appropriate summary statistics, hypothesis test and visualisation and applies them.
it uses \code{\link{map_to_case}},\code{\link{map_to_indicator}},\code{\link{map_to_hypothesis}},\code{\link{map_to_visualisation}}
}
\examples{
plot_crayons()
}
