% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_analysis_plan.R
\name{analysisplan_expand_repeat}
\alias{analysisplan_expand_repeat}
\title{expand an analysis plan with repeat var}
\usage{
analysisplan_expand_repeat(analysisplan, data)
}
\description{
each repetion gets its own analysisplan row
}
