% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{combine_weighting_functions}
\alias{combine_weighting_functions}
\title{Combine weight functions from two sampling frames}
\usage{
combine_weighting_functions(weight_function_1, weight_function_2)
}
\arguments{
\item{weight_function_1}{first weighthing function}

\item{weight_function_2}{second weightng function}
}
\value{
returns a new function that takes a data frame as input returns a vector of weights corresponding to each row in the data frame.
}
\description{
Combine weight functions from two sampling frames
}
