% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_analysis_plan.R
\name{from_analysisplan_map_to_output}
\alias{from_analysisplan_map_to_output}
\title{apply an analysis plan}
\usage{
from_analysisplan_map_to_output(data, analysisplan, weighting = NULL,
  cluster_variable_name = NULL, questionnaire = NULL,
  labeled = FALSE, verbose = TRUE, confidence_level = 0.95)
}
\arguments{
\item{data}{the data set as a data frame (load_data())}

\item{analysisplan}{the analysisplan (load_analysisplan())}

\item{weighting}{optional: the weighting function (use load_samplingframe() and then map_to_weighting())}

\item{cluster_variable_name}{optional: the name of the variable with the cluster IDs}

\item{questionnaire}{optional: the questionnaire (load_questionnaire())}

\item{labeled}{do you want the resuts to display labels rather than xml names ? defaults to false, requires the questionnaire}

\item{verbose}{should progress be printed to the console? (default TRUE, slightly faster if FALSE)}

\item{confidence_level}{the confidence level to be used for confidence intervals (default: 0.95)}
}
\value{
returns a list of hypegrammaR "result" objects (see map_to_result())
}
\description{
Takes all usual hypegrammaR input files plus an analysis plan and maps directly to an output document
}
