% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{guess_to_case}
\alias{guess_to_case}
\title{Guess the  case}
\usage{
guess_to_case(data, hypothesis.type, dependent.var,
  independent.var = NULL, dependent.var.type = NULL,
  independent.var.type = NULL, paired = NULL)
}
\arguments{
\item{data}{missing documentation}
}
\value{
a string that other functions can use to know what analysis case they are dealing with. It has a class "analysis_case" assigned
}
\description{
creates a string that other functions can use to know what analysis case they are dealing with
}
\examples{
map_to_design(data,cluster_variable_name="cluster_id")
}
