% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_tests.R
\name{hypothesis_test_chisquared}
\alias{hypothesis_test_chisquared}
\title{DEPRECIATED
select_one or select_multiple choice now made by map_to_summary_statistic}
\usage{
hypothesis_test_chisquared(dependent.var, independent.var, design)
}
\description{
DEPRECIATED
select_one or select_multiple choice now made by map_to_summary_statistic
}
