% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_tests.R
\name{hypothesis_test_chisquared_select_multiple}
\alias{hypothesis_test_chisquared_select_multiple}
\title{Perform a chi squared test on a select multiple question against a select one question.}
\usage{
hypothesis_test_chisquared_select_multiple(dependent.var,
  dependent.var.sm.cols, independent.var, design, questionnaire = NULL)
}
\arguments{
\item{dependent.var}{string with the column name in `data` of the dependent variable. Should be a 'select multiple'.}

\item{design}{the svy design object created using map_to_design or directly with svydesign}

\item{independen.var}{string with the column name in `data` of the independent variable. Should be a 'select one' with few (<15) categories.}
}
\value{
A list with the results of the test (Chi Squared statistics, p value) or the error message.
}
\description{
Perform a chi squared test on a select multiple question against a select one question.
}
\examples{
hypothesis_test_chisquared_select_one("population_group", "resp_gender", design)
}
