% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_tests.R
\name{hypothesis_test_chisquared_select_one}
\alias{hypothesis_test_chisquared_select_one}
\title{hypothesis_test_chisquared_select_one
Perform a chi squared test on a select one question against another.}
\usage{
hypothesis_test_chisquared_select_one(dependent.var, independent.var,
  design)
}
\arguments{
\item{dependent.var}{string with the column name in `data` of the dependent variable. Should be a 'select one'.}

\item{design}{the svy design object created using map_to_design or directly with svydesign}

\item{independen.var}{string with the column name in `data` of the independent variable. Should be a 'select one' with few (<15) categories.}
}
\value{
A list with the results of the test (Chi Squared statistics, p value) or the error message.
}
\description{
hypothesis_test_chisquared_select_one
Perform a chi squared test on a select one question against another.
}
\examples{
hypothesis_test_chisquared_select_one("population_group", "resp_gender", design)
}
