% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_tests.R
\name{hypothesis_test_t_one_sample}
\alias{hypothesis_test_t_one_sample}
\title{Perform a one sample t test of one numerical variable against  hypothesised value (limit)}
\usage{
hypothesis_test_t_one_sample(dependent.var, independent.var = NULL,
  limit, design)
}
\arguments{
\item{dependent.var}{string with the column name in `data` of the dependent variable. Should be numerical.}

\item{independent.var}{should be null ! For other functions: string with the column name in `data` of the independent variable}

\item{limit}{the value to test the dependent.var against}

\item{design}{the svy design object created using map_to_design or directly with svydesign}
}
\value{
A list with the results of the test (T-value, p value, etc.) or the error message.
}
\description{
Perform a one sample t test of one numerical variable against  hypothesised value (limit)
}
\examples{
\dontrun{hypothesis_test_t_two_sample("males_13_15", 4, design)}
}
