% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_analysisplan}
\alias{load_analysisplan}
\title{Load an analysis plan from a csv file}
\usage{
load_analysisplan(file = NULL, df = NULL)
}
\arguments{
\item{file}{path to a csv file with the analysis plan}

\item{df}{alternative to `file`, you can provide the analysis plan as a data frame}
}
\description{
Load an analysis plan from a csv file
}
