% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_questionnaire}
\alias{load_questionnaire}
\title{Create questionnaire from csv files}
\usage{
load_questionnaire(...)
}
\description{
Create questionnaire from csv files
}
\details{
This enables new functions associated with the questionnaire. It uses load_questionnaire() from the koboquest package.
}
