% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_samplingframe}
\alias{load_samplingframe}
\title{Create weighting from a sampling frame}
\usage{
load_samplingframe(file)
}
\description{
Create weighting from a sampling frame
}
\details{
Create a 'weighter' function from a sampling frame data frame. Uses surveyweights::weighting_fun_from_samplingframe()
}
