% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{map_to_case}
\alias{map_to_case}
\title{Map to case}
\usage{
map_to_case(data, hypothesis.type, dependent.var, independent.var = NULL,
  paired = NULL)
}
\arguments{
\item{data}{missing documentation}
}
\value{
a string that other functions can use to know what analysis case they are dealing with. It has a class "analysis_case" assigned
}
\description{
creates a string that other functions can use to know what analysis case they are dealing with
}
\examples{
map_to_design(data,cluster.var="cluster_id")
}
