% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{map_to_design}
\alias{map_to_design}
\title{Map to Design}
\usage{
map_to_design(data, cluster.var = NULL)
}
\arguments{
\item{cluster.var}{if cluster sampling was used, what's the name of the column in `data` that identifies the cluster?}
}
\value{
a `survey` package design object
}
\description{
creates a `survey` design object from the data
}
\details{
create a `survey` package design object from the data and information on the sampling strategy
}
\examples{
map_to_design(data,cluster.var="cluster_id")
}
