% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_generic_hierarchical_template.R
\name{map_to_generic_hierarchical_html}
\alias{map_to_generic_hierarchical_html}
\title{html from resultlist with results in specified hierarchical order based on analysisplan}
\usage{
map_to_generic_hierarchical_html(resultlist, render_result_with,
  by_analysisplan_columns = c("dependent.var"), by_prefix = c("",
  "subset:", "variable:"), level = 2, questionnaire = NULL,
  label_varnames = TRUE, dir = "./", filename)
}
\arguments{
\item{resultlist}{structure like the output from from_analysisplan_map_to_output:
A list with two items "analysisplan" and "results": The "analysisplan" as a data frame, where each row must match a  result in a list of "results"}

\item{render_result_with}{a function that takes a single result as input and returns an rmarkdown formated string}

\item{by_analysisplan_columns}{vector of strings matching column names of the analysisplan. The first element becomes the main heading, the second element the sub-heading etc.}

\item{by_prefix}{a prefix added at the beginnig of the headline; same length as `by_analysisplan_columns`}

\item{level}{the markdown header level to start with; defaults to 2 which leads to "## heading", i.e. the second header level.}

\item{questionnaire}{optional; the questionnaire (koboquest::load_questionnaire())}

\item{label_varnames}{wether variables names should be labeled in headings}

\item{dir}{the directory in which to save the output file (absolute path or relative to current working directory)}

\item{filename}{the name of the file. must end in '.html'}

\item{type}{the type of report template to use. Currently one of "full", "visual" or "summary"}
}
\description{
html from resultlist with results in specified hierarchical order based on analysisplan
}
