% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_hypothesis_test.R
\name{map_to_hypothesis_test}
\alias{map_to_hypothesis_test}
\title{map to hypothesis test}
\usage{
map_to_hypothesis_test(design, dependent.var, independent.var, case,
  questionnaire = NULL, limit = NULL)
}
\arguments{
\item{case}{a string uniquely identifying the analysis case. output of map_to_case().}
}
\value{
a _function_ that computes the relevant hypothesis test
}
\description{
selects an appropriate hypothesis test function based on the analysis case
}
