% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_labeled.R
\name{map_to_labeled}
\alias{map_to_labeled}
\title{Add labels to results}
\usage{
map_to_labeled(result, questionnaire)
}
\arguments{
\item{result}{hypegrammaR `result` object; output from map_to_result().}

\item{questionnaire}{koboquest `questionnaire` object; output from load_questionnaire()}
}
\value{
same as `result` input, but with all variable values labeled
}
\description{
Add labels to results
}
\details{
if the variable wasn't found in the questionnaire, or the choice wasn't found in the corresponding list of choices, the affected values will remain unchanged.
}
