% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{map_to_questionnaire}
\alias{map_to_questionnaire}
\title{Create questionnaire from csv files}
\usage{
map_to_questionnaire(...)
}
\description{
Create questionnaire from csv files
}
\details{
This enables new functions associated with the questionnaire. It uses load_questionnaire() from the koboquest package.
}
