% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_summary_statistic.R
\name{map_to_summary_statistic}
\alias{map_to_summary_statistic}
\title{Map to summary statistic}
\usage{
map_to_summary_statistic(design, dependent.var, independent.var, case,
  questionnaire = NULL)
}
\arguments{
\item{case}{a string uniquely identifying the analysis case. output of map_to_case().}
}
\value{
a _function_ that computes the relevant summary statistic
}
\description{
selects an appropriate summary statistic function based on the analysis case
}
\examples{
map_to_summary_statistic("group_difference_categorical_categorical")
my_case<- map_to_case( ... )
my_sumstat <- map_to_summary_statistic(my_case)
my_sumstat( ... )
}
