% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_table.R
\name{map_to_table}
\alias{map_to_table}
\title{results as a table}
\usage{
map_to_table(result)
}
\arguments{
\item{result}{a hypegrammaR `result` object produced by map_to_result}
}
\description{
results as a table
}
