% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{map_to_visualisation}
\alias{map_to_visualisation}
\title{map to visualisation}
\usage{
map_to_visualisation(result)
}
\arguments{
\item{result}{a result object containing the summary statistics and hypothesis tests for the case.}
}
\value{
a _function_ that creates the relevant ggplot object
}
\description{
selects an appropriate visualisation function based on the analysis case
}
\examples{
map_to_visualisation("result_var1")
result_var1<- map_to_result( ... )
my_vis_fun <- map_to_visualisation(result_var1)
my_ggplot_obj<-my_vis_fun( ... )
my_ggplot_obj # plots the object
}
