% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_heatmap.R
\name{map_to_visualisation_heatmap}
\alias{map_to_visualisation_heatmap}
\title{Heatmaps from `result` objects}
\usage{
map_to_visualisation_heatmap(result)
}
\arguments{
\item{result}{a hypegrammaR result object (can be made with map_to_result())}
}
\value{
A hypegrammaR visualisation object, which is a list with two elements, 1) a ggplot object and 2) recommended parameters to pass to ggsave.
}
\description{
Heatmaps from `result` objects
}
\details{
to add labels, use `myresult %>% map_to_labeled %>% map_to_result`. The easiest way to save this to a file is with `map_to_file()`. The first elemnent of the returned list is a ggplot object, so you can add/overwrite ggplot elements to it.
}
