% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{map_to_weighting}
\alias{map_to_weighting}
\title{Create weighting from a sampling frame}
\usage{
map_to_weighting(...)
}
\description{
Create weighting from a sampling frame
}
\details{
Create a 'weighter' function from a sampling frame data frame. Uses surveyweights::weighting_fun_from_samplingframe()
}
