% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{mean_with_confints}
\alias{mean_with_confints}
\title{Weighted means with confidence intervals}
\usage{
mean_with_confints(dependent.var, independent.var = NULL, design)
}
\arguments{
\item{dependent.var}{string with the column name in `data` of the dependent variable. Should be a numerical variable.}

\item{independent.var}{should be null ! For other functions: string with the column name in `data` of the independent variable}

\item{design}{the svy design object created using map_to_design or directly with svydesign}
}
\value{
A table in long format of the results, with the column names dependent.var, dependent.var.value (=NA), independent.var (= NA), independent.var.value (= NA), numbers (= mean), se, min and max.
}
\description{
Weighted means with confidence intervals
}
\details{
This function takes the design object and the name of your dependent variable when the latter is a numerical. It calculates the weighted mean for your variable.
}
