% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{mean_with_confints_groups}
\alias{mean_with_confints_groups}
\title{Weighted means with confidence intervals for groups}
\usage{
mean_with_confints_groups(dependent.var, independent.var, design,
  confidence_level = 0.95)
}
\arguments{
\item{dependent.var}{string with the column name in `data` of the dependent variable. Should be a numerical variable.}

\item{independent.var}{string with the column name in `data` of the independent (group) variable. Should be a 'select one'}

\item{design}{the svy design object created using map_to_design or directly with svydesign}

\item{confidence_level}{the confidence level to be used for confidence intervals (default: 0.95)}
}
\value{
A table in long format of the results, with the column names dependent.var, dependent.var.value (=NA), independent.var, independent.var.value, numbers (= mean), se, min and max.
}
\description{
Weighted means with confidence intervals for groups
}
\details{
This function takes the design object and the name of your dependent variable when the latter is a numerical. It calculates the weighted mean for your variable.
}
