% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{percent_with_confints_select_multiple}
\alias{percent_with_confints_select_multiple}
\title{Weighted percentages with confidence intervals for select multiple questions}
\usage{
percent_with_confints_select_multiple(dependent.var, dependent.var.sm.cols,
  design, na.rm = TRUE)
}
\arguments{
\item{dependent.var}{string with the column name in `data` of the dependent variable. Should be a 'select multiple.}

\item{dependent.var.sm.cols}{a vector with the columns indices of the choices for the select multiple question. Can be obtained by calling choices_for_Select_multiple(question.name, data)}

\item{design}{the svy design object created using map_to_design or directly with svydesign}
}
\value{
A table in long format of the results, with the column names dependent.var, dependent.var.value, independent.var (= NA), independent.var.value (= NA), numbers, se, min and max.
}
\description{
Weighted percentages with confidence intervals for select multiple questions
}
\details{
this function takes the design object and the name of your dependent variable when this one is a select multiple. It calculates the weighted percentage for each category.
}
