% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{percent_with_confints_select_multiple_groups}
\alias{percent_with_confints_select_multiple_groups}
\title{Weighted percentages with confidence intervals for groups (select multiple questions)}
\usage{
percent_with_confints_select_multiple_groups(dependent.var,
  dependent.var.sm.cols, independent.var, design, na.rm = TRUE,
  confidence_level = 0.95)
}
\arguments{
\item{dependent.var}{string with the column name in `data` of the dependent variable. Should be a 'select multiple.}

\item{dependent.var.sm.cols}{a vector with the columns indices of the choices for the select multiple question. Can be obtained by calling choices_for_Select_multiple(question.name, data)}

\item{independent.var}{string with the column name in `data` of the independent (group) variable. Should be a 'select one'}

\item{design}{the svy design object created using map_to_design or directly with svydesign}

\item{confidence_level}{the confidence level to be used for confidence intervals (default: 0.95)}
}
\value{
A table in long format of the results, with the column names dependent.var, dependent.var.value, independent.var (= NA), independent.var.value (= NA), numbers, se, min and max.
}
\description{
Weighted percentages with confidence intervals for groups (select multiple questions)
}
\details{
this function takes the design object and the name of your dependent variable when this one is a select multiple. It calculates the weighted percentage for each category.
}
