% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_colours.R
\name{reach_style_color_darkgreys}
\alias{reach_style_color_darkgreys}
\title{Reach brand dark grey triples}
\usage{
reach_style_color_darkgreys()
}
\description{
Reach brand dark grey triples
}
\examples{



}
\seealso{
\code{\link{function_name}}
}
