% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{read.csv.auto.sep}
\alias{read.csv.auto.sep}
\title{loading function with automatic default}
\usage{
read.csv.auto.sep(file, stringsAsFactors = F, ...)
}
\arguments{
\item{file}{path to a csv file with the assessment data}
}
\value{
the data from the csv files as data frame. Column header symbols are changed to lowercase alphanumeric and underscore; everything else is converted to a "."
}
\description{
loading function with automatic default
}
\details{
the file is loaded with stringsAsFactors = F and with column names in alphanumeric lowercase
}
