% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_result_utils.R
\name{resultlist_recursive_markdown}
\alias{resultlist_recursive_markdown}
\title{Rmarkdown from resultlist in specified hierarchical order}
\usage{
resultlist_recursive_markdown(resultlist,
  by_analysisplan_columns = c("dependent.var"), by_prefix = c("",
  "subset:", "variable:"), level = 2, render_result_with,
  questionnaire = NULL, label_varnames = TRUE)
}
\arguments{
\item{resultlist}{structure like the output from from_analysisplan_map_to_output:
A list with two items "analysisplan" and "results": The "analysisplan" as a data frame, where each row must match a  result in a list of "results"}

\item{by_analysisplan_columns}{vector of strings matching column names of the analysisplan. The first element becomes the main heading, the second element the sub-heading etc.}

\item{by_prefix}{a prefix added at the beginnig of the headline; same length as `by_analysisplan_columns`}

\item{level}{the markdown header level to start with; defaults to 2 which leads to "## heading", i.e. the second header level.}

\item{render_result_with}{a function that takes a single result as input and returns an rmarkdown formated string}

\item{questionnaire}{optional; the questionnaire (koboquest::load_questionnaire())}

\item{label_varnames}{wether variables names should be labeled in headings}
}
\description{
Rmarkdown from resultlist in specified hierarchical order
}
