% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_subset.R, R/template_result_utils.R
\name{results_subset}
\alias{results_subset}
\title{subset a list of results based on analysis parameters}
\usage{
results_subset(results, repeat.vars = NULL, repeat.var.values = NULL,
  dependent.vars = NULL, logical = NULL)

results_subset(results, repeat.vars = NULL, repeat.var.values = NULL,
  dependent.vars = NULL, logical = NULL)
}
\arguments{
\item{results}{list of results (output from `from_analysisplan_map_to_output()`)}

\item{repeat.vars}{optional: vector of character strings: keeps only results where repeat.var in this list}

\item{repeat.var.values}{optional: vector of character strings: keeps only results where repeat.var.vaues in this list}

\item{dependent.vars}{optional: vector of character strings: keeps only results where dependent.var in this list}

\item{logical}{optional: subset by a logical vector (same length as list of results)}

\item{results}{list of results (output from `from_analysisplan_map_to_output()`)}

\item{repeat.vars}{optional: vector of character strings: keeps only results where repeat.var in this list}

\item{repeat.var.values}{optional: vector of character strings: keeps only results where repeat.var.vaues in this list}

\item{dependent.vars}{optional: vector of character strings: keeps only results where dependent.var in this list}

\item{logical}{optional: subset by a logical vector (same length as list of results)}
}
\value{
a resultlist in same format as from_analysisplan_map_to_output() only including those results with matching analysis parameters

a resultlist in same format as from_analysisplan_map_to_output() only including those results with matching analysis parameters
}
\description{
subset a list of results based on analysis parameters

subset a list of results based on analysis parameters
}
\details{
if multiple parameters are given to subset by, only those are kept where all conditions apply

if multiple parameters are given to subset by, only those are kept where all conditions apply
}
