% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{summary_statistic_mode_select_one}
\alias{summary_statistic_mode_select_one}
\title{Weighted means with confidence intervals for groups}
\usage{
summary_statistic_mode_select_one(dependent.var, independent.var, design)
}
\arguments{
\item{dependent.var}{string with the column name in `data` of the dependent variable. Should be a select_one or a select_multiple.}

\item{independent.var}{string with the column name in `data` of the independent (group) variable. Should be a 'select one'}

\item{design}{the svy design object created using map_to_design or directly with svydesign}
}
\value{
A table in long format of the results, with the column names dependent.var, dependent.var.value (=NA), independent.var, independent.var.value, numbers (= mean), se, min and max.
}
\description{
Weighted means with confidence intervals for groups
}
\details{
This function takes the design object and the name of your dependent variable, and returns the most frequent answer for each category in independent.var
}
