% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_data}
\alias{load_data}
\title{load asessment data}
\usage{
load_data(file)
}
\arguments{
\item{file}{path to a csv file with the assessment data}
}
\value{
the data from the csv files as data frame. Column header symbols are changed to lowercase alphanumeric and underscore; everything else is converted to a "."
}
\description{
load asessment data
}
\details{
the data _must_ be in standard kobo format with xml style headers.
}
