% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_samplingframe}
\alias{load_samplingframe}
\title{Load a sampling frame from csv}
\usage{
load_samplingframe(file)
}
\arguments{
\item{file}{the path and name of the sampling frame csv file to load.}
}
\description{
Load a sampling frame from csv
}
\details{
function loads the sampling frame and can be used to make weights ith map_to_weighting()
}
\examples{
sf <- load_samplingframe("./somefolder/samplingframe.csv")

}
