% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_samplingframe}
\alias{load_samplingframe}
\title{Load a sampling frame from csv}
\usage{
load_samplingframe(file)
}
\arguments{
\item{file}{the path and name of the sampling frame csv file to load.}
}
\description{
Load a sampling frame from csv
}
\examples{

# load the sampling frame:
sf <- load_samplingframe("./somefolder/samplingframe.csv")

# it can be used to make weights with map_to_weighting()


}
