% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_case.R
\name{map_to_case}
\alias{map_to_case}
\title{Map to case}
\usage{
map_to_case(hypothesis.type, dependent.var.type = NULL,
  independent.var.type = NULL)
}
\arguments{
\item{hypothesis.type}{The hypothesis type. Must be one of "group_difference" or "direct_reporting".}

\item{dependent.var.type}{The type of the dependent variable as a string. must be either "numerical" or "categorical"}

\item{independent.var.type}{The type of the independent variable as a string. must be either "numerical" or "categorical"}
}
\value{
a string that other functions can use to know what analysis case they are dealing with. It has a class "analysis_case" assigned
}
\description{
creates a string that other functions can use to know what analysis case they are dealing with
}
\examples{
\dontrun{map_to_case("group_difference","categorical","categorical")}
}
