% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{map_to_design}
\alias{map_to_design}
\title{Map to Design}
\usage{
map_to_design(data, cluster_variable_name = NULL,
  weighting_function = NULL)
}
\arguments{
\item{data}{the dataset as a sampling frame. Must match the sampling frame provided to create the `weighting_function` produced with `map_to_weighting()`}

\item{weighting_function}{if cluster sampling was used, what's the name of the column in `data` that identifies the cluster?}
}
\value{
a `survey` package design object
}
\description{
creates a `survey` design object from the data
}
\details{
create a `survey` package design object from the data and information on the sampling strategy
}
\examples{
\dontrun{map_to_design(data,cluster_variable_name="cluster_id")}
}
