% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_hypothesis_test.R
\name{map_to_hypothesis_test}
\alias{map_to_hypothesis_test}
\title{map to hypothesis test}
\usage{
map_to_hypothesis_test(design, dependent.var, independent.var, case,
  questionnaire = NULL)
}
\arguments{
\item{case}{a string uniquely identifying the analysis case. output of \code{\link{map_to_case}}. To list valid case strings use \link{\code{list_all_cases}}}
}
\value{
a _function_ that computes the relevant hypothesis test
}
\description{
selects an appropriate hypothesis test function based on the analysis case
}
\examples{
map_to_summary_statistic("group_difference_categorical_categorical")
my_case<- map_to_case( ... )
my_hyptest <- map_to_hypothesis_test(my_case)
my_hyptest( ... )
}
