% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{map_to_summary_statistic}
\alias{map_to_summary_statistic}
\title{Map to summary statistic}
\usage{
map_to_summary_statistic(case)
}
\arguments{
\item{case}{a string uniquely identifying the analysis case. output of \link{\code{map_to_case}}. To list valid case strings use \link{\code{list_all_cases}}}
}
\value{
a _function_ that computes the relevant summary statistic
}
\description{
selects an appropriate summary statistic function based on the analysis case
}
\examples{
map_to_summary_statistic("group_difference_categorical_categorical")
my_case<- map_to_case( ... )
my_sumstat <- map_to_summary_statistic(my_case)
my_sumstat( ... )
}
