% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{map_to_visualisation}
\alias{map_to_visualisation}
\title{map to visualisation}
\usage{
map_to_visualisation(case)
}
\arguments{
\item{case}{a string uniquely identifying the analysis case. output of \link{\code{map_to_case}}. To list valid case strings use \link{\code{list_all_cases}}}
}
\value{
a _function_ that creates the relevant ggplot object
}
\description{
selects an appropriate visualisation function based on the analysis case
}
\examples{
map_to_visualisation("group_difference_categorical_categorical")
my_case<- map_to_case( ... )
my_vis_fun <- map_to_visualisation(my_case)
my_ggplot_obj<-my_vis_fun( ... )
my_ggplot_obj # plots the object
}
